requirements.txt
streamlit
openai
sentence-transformers
chromadb
faiss-cpu
python-dotenv
